🔨 extract shared InlineLabeledDropdown component
